[Improve][Test] Reduce the frequency of test failures of `testApplyResourceWithRandomResult`
