--- seatunnel-engine/seatunnel-engine-server/src/test/java/org/apache/seatunnel/engine/server/resourcemanager/ResourceManagerFunctionTest.java
@@ -1,56 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.seatunnel.engine.server.resourcemanager;
-
-import org.apache.seatunnel.engine.server.AbstractSeaTunnelServerTest;
-import org.apache.seatunnel.engine.server.resourcemanager.resource.ResourceProfile;
-import org.apache.seatunnel.engine.server.resourcemanager.resource.SlotProfile;
-
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-
-import com.hazelcast.cluster.Address;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Set;
-import java.util.concurrent.ExecutionException;
-import java.util.stream.Collectors;
-
-public class ResourceManagerFunctionTest
-        extends AbstractSeaTunnelServerTest<ResourceManagerFunctionTest> {
-
-    @Test
-    public void testApplyResourceWithRandomResult()
-            throws ExecutionException, InterruptedException {
-        FakeResourceManager resourceManager = new FakeResourceManager(nodeEngine);
-
-        List<ResourceProfile> resourceProfiles = new ArrayList<>();
-        resourceProfiles.add(new ResourceProfile());
-        resourceProfiles.add(new ResourceProfile());
-        resourceProfiles.add(new ResourceProfile());
-        resourceProfiles.add(new ResourceProfile());
-        resourceProfiles.add(new ResourceProfile());
-        List<SlotProfile> slotProfiles = resourceManager.applyResources(1L, resourceProfiles).get();
-        Assertions.assertEquals(slotProfiles.size(), 5);
-
-        Set<Address> addresses =
-                slotProfiles.stream().map(SlotProfile::getWorker).collect(Collectors.toSet());
-        Assertions.assertTrue(addresses.size() > 1);
-    }
-}
